Create README.md
# SPOTlight: Seeded NMF regression to Deconvolute Spatial Transcriptomics Spots with Single-Cell Transcriptomes

This repository contains all the scripts, notebooks, and reports used to reproduce the analysis of our paper SPOTlight: Seeded NMF regression to Deconvolute Spatial Transcriptomics Spots with Single-Cell Transcriptomes. Here we describe how to access the data, the most important packages and versions used, and how to navigate the directories and files in the repository.